# openapi/spec.yaml
openapi: 3.0.3
info:
  title: AI Support Platform API
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
  /healthz:
    get:
      operationId: getHealth
      summary: Health check
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                required:
                  - status
                properties:
                  status:
                    type: string
  /v1/auth/login:
    post:
      operationId: login
      summary: Login and issue a JWT access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Validation failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
  /v1/tickets/{ticketID}:
    get:
      operationId: getTicket
      summary: Fetch a ticket by ID
      security:
        - bearerAuth: []
      parameters:
        - name: ticketID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Ticket found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketResponse"
        "400":
          description: Validation failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: Ticket not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
    patch:
      operationId: updateTicket
      summary: Update an existing ticket
      security:
        - bearerAuth: []
      parameters:
        - name: ticketID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTicketRequest"
      responses:
        "200":
          description: Ticket updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketResponse"
        "400":
          description: Validation failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: Ticket not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
  /v1/tickets:
    get:
      operationId: listTickets
      summary: List tickets for the authenticated tenant
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Tickets found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TicketResponse"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
    post:
      operationId: createTicket
      summary: Create a new ticket
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTicketRequest"
      responses:
        "201":
          description: Ticket created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketResponse"
        "400":
          description: Validation failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginRequest:
      type: object
      required:
        - email
        - password
        - tenant_id
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        tenant_id:
          type: string
          format: uuid
    LoginResponse:
      type: object
      required:
        - access_token
        - token_type
        - user_id
        - tenant_id
        - role
      properties:
        access_token:
          type: string
        token_type:
          type: string
        user_id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        role:
          type: string
    CreateTicketRequest:
      type: object
      required:
        - subject
      properties:
        subject:
          type: string
    UpdateTicketRequest:
      type: object
      properties:
        subject:
          type: string
        status:
          type: string
          enum:
            - open
            - closed
    TicketResponse:
      type: object
      required:
        - id
        - tenant_id
        - subject
        - status
        - created_by_user_id
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        subject:
          type: string
        status:
          type: string
        created_by_user_id:
          type: string
          format: uuid
        assigned_to_user_id:
          type: string
          format: uuid
          nullable: true
    Problem:
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        request_id:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string